Rename files from rgb_to_greyscale to color_to_greyscale
