+++ src/modules/cpu/host_tensor_data_exchange_operations.hpp
@@ -25,6 +25,6 @@ THE SOFTWARE.
 
 #include "kernel/copy.hpp"
 #include "kernel/swap_channels.hpp"
-#include "kernel/rgb_to_greyscale.hpp"
+#include "kernel/color_to_greyscale.hpp"
 
 #endif // HOST_TENSOR_DATA_EXCHANGEE_OPERATIONS_HPP
