+++ src/modules/cpu/kernel/color_to_greyscale.hpp
