+++ src/modules/hip/hip_tensor_data_exchange_operations.hpp
@@ -25,6 +25,6 @@ THE SOFTWARE.
 
 #include "kernel/copy.hpp"
 #include "kernel/swap_channels.hpp"
-#include "kernel/rgb_to_greyscale.hpp"
+#include "kernel/color_to_greyscale.hpp"
 
 #endif // HIP_TENSOR_DATA_EXCHANGE_OPERATIONS_HPP
