+++ src/modules/hip/kernel/color_to_greyscale.hpp
